requirements.txt
fastapi
uvicorn[standard]
python-dotenv
PyJWT
sqlmodel
pydantic[email]
sqlalchemy
passlib
bcrypt
python-multipart
pandas
python-slugify
